huggingface_space/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# RAG and AI dependencies
qdrant-client>=1.6.0
cohere>=4.30
aiohttp>=3.8.0
tenacity>=8.2.0
numpy>=1.20.0
python-dotenv>=1.0.0

# Text processing
markdown>=3.5.0
python-docx>=0.8.11
PyPDF2>=3.0.0

# Data validation
pydantic>=2.0.0
